+++ tests/run_mongo_with_tests.sh
@@ -1,0 +1,9 @@
+#!/bin/sh
+# run tests with mongodb running
+docker pull mongo
+ID=`docker run -d --network=host mongo`
+
+pytest .
+
+docker stop $ID
+docker rm $ID